+++ src/main/java/roomescape/dto/ReservationRequest.java
@@ -1,11 +1,11 @@
 package roomescape.dto;
 
 import java.time.LocalDate;
-import roomescape.domain.LoginMember;
-import roomescape.domain.Member;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.Theme;
+
+import roomescape.entity.Member;
+import roomescape.entity.Reservation;
+import roomescape.entity.ReservationTime;
+import roomescape.entity.Theme;
 
 public record ReservationRequest(LocalDate date, long timeId, long themeId) {
     public Reservation toReservation(Member member, ReservationTime reservationTime, Theme theme) {
