+++ src/main/java/roomescape/dto/ReservationTimeResponse.java
@@ -1,7 +1,7 @@
 package roomescape.dto;
 
 import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
+import roomescape.entity.ReservationTime;
 
 public record ReservationTimeResponse(long id, LocalTime startAt) {
     public static ReservationTimeResponse from(ReservationTime reservationTime) {
