+++ src/main/java/roomescape/dto/ThemeRequest.java
@@ -1,6 +1,6 @@
 package roomescape.dto;
 
-import roomescape.domain.Theme;
+import roomescape.entity.Theme;
 
 public record ThemeRequest(String name, String description, String thumbnail) {
     public Theme toTheme() {
