+++ src/main/java/roomescape/repository/MemberRepository.java
@@ -1,10 +1,8 @@
 package roomescape.repository;
 
-import jakarta.persistence.Id;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import roomescape.domain.Member;
-import roomescape.domain.Members;
+import roomescape.entity.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByEmail(String email);
