+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -1,10 +1,9 @@
 package roomescape.repository;
 
 import java.time.LocalTime;
-import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.ReservationTimes;
+import roomescape.entity.ReservationTime;
 
 public interface ReservationTimeRepository extends JpaRepository<ReservationTime, Long> {
     boolean existsByStartAt(LocalTime startAt);
