+++ src/main/java/roomescape/repository/ThemeRepository.java
@@ -1,10 +1,8 @@
 package roomescape.repository;
 
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
-import roomescape.domain.Duration;
-import roomescape.domain.Theme;
-import roomescape.domain.Themes;
+
+import roomescape.entity.Theme;
 
 public interface ThemeRepository extends JpaRepository<Theme, Long> {
 }
