+++ src/main/java/roomescape/service/ReservationService.java
@@ -11,11 +11,11 @@ import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.LoginMember;
-import roomescape.domain.Member;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
+import roomescape.entity.Member;
+import roomescape.entity.Reservation;
+import roomescape.entity.ReservationTime;
 import roomescape.domain.Reservations;
-import roomescape.domain.Theme;
+import roomescape.entity.Theme;
 import roomescape.dto.AdminReservationRequest;
 import roomescape.dto.ReservationDetailResponse;
 import roomescape.dto.ReservationRequest;
