+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -7,10 +7,10 @@ import static roomescape.exception.ExceptionType.NOT_FOUND_THEME;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.domain.ReservationTime;
+import roomescape.entity.ReservationTime;
 import roomescape.domain.ReservationTimes;
 import roomescape.domain.Reservations;
-import roomescape.domain.Theme;
+import roomescape.entity.Theme;
 import roomescape.dto.AvailableTimeResponse;
 import roomescape.dto.ReservationTimeRequest;
 import roomescape.dto.ReservationTimeResponse;
