+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,14 +3,11 @@ package roomescape.service;
 import static roomescape.exception.ExceptionType.DELETE_USED_THEME;
 import static roomescape.exception.ExceptionType.DUPLICATE_THEME;
 
-import java.util.Comparator;
 import java.util.List;
-import java.util.Map.Entry;
-import java.util.function.Function;
-import java.util.stream.Collectors;
+
 import org.springframework.stereotype.Service;
 import roomescape.domain.Duration;
-import roomescape.domain.Theme;
+import roomescape.entity.Theme;
 import roomescape.domain.Themes;
 import roomescape.dto.ThemeRequest;
 import roomescape.dto.ThemeResponse;
