+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -8,6 +8,8 @@ import static roomescape.exception.ExceptionType.EMPTY_TIME;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import roomescape.entity.ReservationTime;
 import roomescape.exception.RoomescapeException;
 
 class ReservationTimeTest {
