+++ src/test/java/roomescape/domain/ThemeTest.java
@@ -10,6 +10,8 @@ import static roomescape.exception.ExceptionType.EMPTY_THUMBNAIL;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import roomescape.entity.Theme;
 import roomescape.exception.RoomescapeException;
 
 class ThemeTest {
