+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -19,10 +19,10 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 import roomescape.Fixture;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
+import roomescape.entity.Reservation;
+import roomescape.entity.ReservationTime;
 import roomescape.domain.ReservationTimes;
-import roomescape.domain.Theme;
+import roomescape.entity.Theme;
 import roomescape.dto.AvailableTimeResponse;
 import roomescape.dto.ReservationTimeRequest;
 import roomescape.dto.ReservationTimeResponse;
